+++ plugins/psslurm/src/psslurm/psslurmjob.h
@@ -38,7 +38,7 @@ typedef struct {
     PStask_ID_t forwarderTID;
     PStask_t *forwarder;
     PStask_group_t childGroup;
-    uint16_t childRank;
+    int32_t childRank;
     int exitCode;
     struct list_head list;
 } PS_Tasks_t;
@@ -355,7 +355,7 @@ void clearAllocList();
 
 PS_Tasks_t *addTask(struct list_head *list, PStask_ID_t childTID,
 			PStask_ID_t forwarderTID, PStask_t *forwarder,
-			PStask_group_t childGroup, uint16_t rank);
+			PStask_group_t childGroup, int32_t rank);
 void signalTasks(uid_t uid, PS_Tasks_t *tasks, int signal, int32_t group);
 PS_Tasks_t *findTaskByRank(struct list_head *taskList, uint16_t rank);
 PS_Tasks_t *findTaskByForwarder(struct list_head *taskList, PStask_ID_t fwTID);
